Add budget recommendation service and refactor cashback to BigDecimal

+++ src/main/java/com/julianmejiac/walletoptimizer/dto/CardRecommendationDTO.java
@@ -1,7 +1,9 @@
 package com.julianmejiac.walletoptimizer.dto;
 
+import java.math.BigDecimal;
+
 public record CardRecommendationDTO (
         String cardName,
-        double cashbackPercent
+        BigDecimal cashbackPercent
 )
 {}
